till about
The Portfolio Website project is a showcase of my skills, experience, and achievements as a web developer. It serves as an online resume and professional portfolio, providing visitors with an overview of my work, projects, and expertise in web development. The website is designed to be visually appealing, easy to navigate, and responsive across various devices.